Feature: security module (#324)
* Move all security data to enigne security-api

* Changes at packages name

* Change dependings at sbt

* Fix tests

* Add provided lazy logging

* Changes at package name.. Add sbt dependencies cdn
