+++ engine/security/src/main/scala/pl/touk/nussknacker/ui/security/api/GlobalPermission.scala
