+++ engine/security/src/main/scala/pl/touk/nussknacker/ui/security/api/NussknackerInternalUser.scala
@@ -1,0 +1,3 @@
+package pl.touk.nussknacker.ui.security.api
+
+object NussknackerInternalUser extends AdminUser("Nussknacker")
