+++ engine/security/src/test/scala/pl/touk/nussknacker/ui/security/api/LoggedUserTest.scala
