+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/initialization/Initialization.scala
@@ -18,8 +18,7 @@ import pl.touk.nussknacker.ui.db.entity.EnvironmentsEntityData
 import pl.touk.nussknacker.ui.process.migrate.ProcessModelMigrator
 import pl.touk.nussknacker.ui.process.repository.WriteProcessRepository.UpdateProcessAction
 import pl.touk.nussknacker.ui.process.repository._
-import pl.touk.nussknacker.ui.security.NussknackerInternalUser
-import pl.touk.nussknacker.ui.security.api.{LoggedUser, Permission}
+import pl.touk.nussknacker.ui.security.api.{LoggedUser, NussknackerInternalUser, Permission}
 import slick.dbio.DBIOAction
 import slick.jdbc.JdbcProfile
