--- ui/server/src/main/scala/pl/touk/nussknacker/ui/security/NussknackerInternalUser.scala
@@ -1,5 +1,0 @@
-package pl.touk.nussknacker.ui.security
-
-import pl.touk.nussknacker.ui.security.api.AdminUser
-
-object NussknackerInternalUser extends AdminUser("Nussknacker")
