--- ui/server/src/test/scala/pl/touk/nussknacker/ui/security/ouath2/OAuth2ServiceProviderSpec.scala
@@ -1,10 +1,0 @@
-package pl.touk.nussknacker.ui.security.ouath2
-
-import org.scalatest.{FlatSpec, Matchers}
-import pl.touk.nussknacker.ui.security.oauth2.{DefaultOAuth2Service, OAuth2ServiceProvider}
-
-class OAuth2ServiceProviderSpec extends FlatSpec with Matchers {
-  it should "return default OAuth2 service" in {
-    OAuth2ServiceProvider(ExampleOAuth2ServiceFactory.testConfig, this.getClass.getClassLoader) shouldBe a[DefaultOAuth2Service]
-  }
-}
